Move parse predicates with error handling to separate file
